requirements.txt
fastapi
uvicorn
openai
transformers
torch
pillow